[rails 4 upgrade, upgrade paperclip]
Remove Imagescalr dependency. …
On Jun 28, 2012 the devs made a decision to swap out Paperclip's default
image manipulation library (Imagemagick), for a javas library named imgscalr.  This is good
in that the imgscalr JAR can be stored in the codebase and, it will run wherever Java is installed.
However, with newer versions of Paperclip, the "PaperclipExtensions" (monkeypatch) that they
wrote to modify Paperclip to work with imgscalr has gone out of date and will need to be
rewritten.

I'd like to re-examine the necessity of using this Java library and consider reverting to the
much more standard path of using Imagemagick.

-----

Newer versions of Paperclip have … changed behavior around validating MIME types, and now
require some form of validation to be present.  I dug into an "example production system"
(10.0.0.204) to look at the extensions of the files uploaded. Then I based these MIME type
 validation configurations off of that research.